refactored project structure to remove dependencies between projects, added DI container integration for Linq2Db project
