+++ Code/src/Light.GuardClauses.cs
