+++ Code/src/Synnotech.Migrations.Core/Synnotech.Migrations.Core.csproj
@@ -4,4 +4,7 @@
 	<PropertyGroup>
 		<DocumentationFile>bin/$(Configuration)/$(TargetFramework)/$(AssemblyName).xml</DocumentationFile>
 	</PropertyGroup>
+	<ItemGroup>
+	  <Compile Include="..\Light.GuardClauses.cs" Link="Light.GuardClauses.cs" />
+	</ItemGroup>
 </Project>
