+++ Code/src/Synnotech.Migrations.Linq2Db/Synnotech.Migrations.Linq2Db.csproj
@@ -1,12 +1,13 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
     <PropertyGroup>
 		<DocumentationFile>bin\$(Configuration)\$(TargetFramework)\$(AssemblyName).xml</DocumentationFile>
 	</PropertyGroup>
     <ItemGroup>
+      <Compile Include="..\Light.GuardClauses.cs" Link="Light.GuardClauses.cs" />
+    </ItemGroup>
+    <ItemGroup>
 		<ProjectReference Include="..\Synnotech.Migrations.Core\Synnotech.Migrations.Core.csproj" />
         <PackageReference Include="linq2db" Version="3.3.0" />
+        <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="5.0.0" />
 	</ItemGroup>
-    <ItemGroup>
-        <Compile Include="..\Synnotech.Migrations.Core\Light.GuardClauses\Light.GuardClauses.cs" Link="Light.GuardClauses\Light.GuardClauses.cs" />
-    </ItemGroup>
 </Project>
