+++ Code/src/Synnotech.Migrations.RavenDB/Synnotech.Migrations.RavenDB.csproj
@@ -3,11 +3,11 @@
 		<DocumentationFile>bin\$(Configuration)\$(TargetFramework)\$(AssemblyName).xml</DocumentationFile>
 	</PropertyGroup>
 	<ItemGroup>
+	  <Compile Include="..\Light.GuardClauses.cs" Link="Light.GuardClauses.cs" />
+	</ItemGroup>
+	<ItemGroup>
 		<ProjectReference Include="..\Synnotech.Migrations.Core\Synnotech.Migrations.Core.csproj" />
 		<PackageReference Include="RavenDB.Client" Version="5.0.0" />
 		<PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="5.0.0" />
-	</ItemGroup>
-	<ItemGroup>
-		<Compile Include="..\Synnotech.Migrations.Core\Light.GuardClauses\Light.GuardClauses.cs" Link="Light.GuardClauses\Light.GuardClauses.cs" />
 	</ItemGroup>
 </Project>
