+++ Code/tests/Synnotech.Migrations.Linq2Db.Tests/TextVersions/Linq2DbMigrationEngineTests.cs
@@ -1,0 +1,7 @@
+﻿namespace Synnotech.Migrations.Linq2Db.Tests.TextVersions
+{
+    public class Linq2DbMigrationEngineTests
+    {
+
+    }
+}
