--- Code/tests/Synnotech.Migrations.RavenDB.Tests/AssemblyInfos.cs
@@ -1,3 +1,0 @@
-﻿using Xunit;
-
-[assembly:CollectionBehavior(DisableTestParallelization = true)]
