+++ Code/tests/Synnotech.Migrations.RavenDB.Tests/Synnotech.Migrations.RavenDB.Tests.csproj
@@ -1,18 +1,18 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
-	<ItemGroup>
-		<ProjectReference Include="..\..\src\Synnotech.Migrations.RavenDB\Synnotech.Migrations.RavenDB.csproj" />
-		<PackageReference Include="RavenDB.TestDriver" Version="5.1.5" />
-		<PackageReference Include="Xunit.SkippableFact" Version="1.4.13" />
-		<PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="5.0.0" />
-		<PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="5.0.0" />
-		<PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="5.0.1" />
-	</ItemGroup>
-	<ItemGroup>
-	  <None Update="testsettings.Development.json" Condition="Exists('testsettings.Development.json')">
-	    <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-	  </None>
-	  <None Update="testsettings.json">
-	    <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-	  </None>
-	</ItemGroup>
+    <ItemGroup>
+        <ProjectReference Include="..\..\src\Synnotech.Migrations.RavenDB\Synnotech.Migrations.RavenDB.csproj" />
+        <PackageReference Include="RavenDB.TestDriver" Version="5.1.5" />
+    </ItemGroup>
+    <ItemGroup>
+        <None Update="testsettings.Development.json" Condition="Exists('testsettings.Development.json')">
+            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+        </None>
+        <None Update="testsettings.json">
+            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+        </None>
+        <Compile Include="..\AssemblyInfos.cs">
+            <Link>AssemblyInfos.cs</Link>
+        </Compile>
+        <Compile Include="..\TestSettings.cs" Link="TestSettings.cs" />
+    </ItemGroup>
 </Project>
